Verify email addresses
Verify the email addresses in the people database before doing anything.
It's not a thorough check, since most SMTP servers won't honestly tell you
if a given address is correct, but it's better than nothing.

Also set the envelope's From address to the sender's email address,
instead of Santa's, so that any bounces should return to the sender.

Signed-off-by: Timur Tabi <timur@tabi.org>