Don't crash if the lastyear file is missing
The lastyear file is named <year>.txt, where <year> is last year.  This
file contains the list of matches from last year.  We keep track of this
so that we don't pick two people to be each other's Secret Santa for two
years in a row.

We weren't catching the exception if the lastyear file was missing, which
is normal the first time you use santa.py.

Signed-off-by: Timur Tabi <timur@tabi.org>